+++ custom_components/imou_life/const.py
@@ -3,7 +3,7 @@
 DOMAIN = "imou_life"
 PLATFORMS = ["switch", "sensor", "binary_sensor"]
 
-# Definitions
+# Configuration definitions
 CONF_DEVICE_NAME = "device_name"
 CONF_APP_ID = "app_id"
 CONF_APP_SECRET = "app_secret"
@@ -14,5 +14,32 @@ OPTION_SCAN_INTERVAL = "scan_interval"
 OPTION_API_URL = "api_url"
 OPTION_API_TIMEOUT = "api_timeout"
 
-# Default values
+# Scan interval
 DEFAULT_SCAN_INTERVAL = 15 * 60
+
+# switches which are enabled by default
+ENABLED_SWITCHES = [
+    "motionDetect",
+    "headerDetect",
+    "abAlarmSound",
+    "breathingLight",
+    "closeCamera",
+    "linkDevAlarm",
+    "whiteLight",
+    "smartTrack",
+]
+
+# icons of the sensors
+SENSOR_ICONS = {
+    "__default__": "mdi:bookmark",
+    "online": "mdi:lan-connect",
+    "lastAlarm": "mdi:timer",
+    "motionDetect": "mdi:motion-sensor",
+    "headerDetect": "mdi:human",
+    "abAlarmSound": "mdi:account-voice",
+    "breathingLight": "mdi:television-ambient-light",
+    "closeCamera": "mdi:sleep",
+    "linkDevAlarm": "mdi:bell",
+    "whiteLight": "mdi:light-flood-down",
+    "smartTrack": "mdi:radar",
+}