+++ custom_components/imou_life/switch.py
@@ -5,10 +5,9 @@ import logging
 from homeassistant.components.switch import SwitchEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
-from imouapi.exceptions import ImouException
 
-from .const import DOMAIN
+from .const import DOMAIN, ENABLED_SWITCHES
+from .entity import ImouEntity
 
 _LOGGER: logging.Logger = logging.getLogger(__package__)
 
@@ -21,7 +20,7 @@ async def async_setup_entry(
     coordinator = hass.data[DOMAIN][entry.entry_id]
     device = coordinator.device
     sensors = []
-    for sensor_instance in coordinator.device.get_sensors("switch"):
+    for sensor_instance in coordinator.device.get_sensors_by_platform("switch"):
         sensor = ImouSwitch(coordinator, entry, sensor_instance)
         sensors.append(sensor)
         _LOGGER.debug(
@@ -30,37 +29,18 @@ async def async_setup_entry(
     async_add_devices(sensors)
 
 
-class ImouSwitch(CoordinatorEntity, SwitchEntity):
+class ImouSwitch(ImouEntity, SwitchEntity):
     """imou switch class."""
 
-    def __init__(self, coordinator, config_entry, sensor_instance):
-        """Initialize."""
-        super().__init__(coordinator)
-        self.config_entry = config_entry
-        self.device = coordinator.device
-        self.sensor_instance = sensor_instance
+    @property
+    def entity_registry_enabled_default(self) -> bool:
+        """If the entity is enabled by default."""
+        return self.sensor_instance.get_name() in ENABLED_SWITCHES
 
     @property
-    def unique_id(self):
-        """Return a unique ID to use for this entity."""
-        return self.config_entry.entry_id + "_" + self.sensor_instance.get_name()
-
-    @property
-    def device_info(self):
-        """Return device information."""
-        return {
-            "identifiers": {(DOMAIN, self.config_entry.entry_id)},
-            "name": self.device.get_name(),
-            "model": self.device.get_model(),
-            "manufacturer": self.device.get_manufacturer(),
-            "sw_version": self.device.get_firmware(),
-            "hw_version": self.device.get_device_id(),
-        }
-
-    @property
-    def available(self) -> bool:
-        """Device available."""
-        return self.coordinator.device.is_online()
+    def is_on(self):
+        """Return true if the switch is on."""
+        return self.sensor_instance.is_on()
 
     async def async_turn_on(self, **kwargs):  # pylint: disable=unused-argument
         """Turn on the switch."""
@@ -85,36 +65,3 @@ class ImouSwitch(CoordinatorEntity, SwitchEntity):
             self.device.get_name(),
             self.sensor_instance.get_description(),
         )
-
-    @property
-    def name(self):
-        """Return the name of the entity."""
-        return f"{self.device.get_name()} {self.sensor_instance.get_description()}"
-
-    @property
-    def icon(self):
-        """Return the icon of this entity."""
-        return self.sensor_instance.get_icon()
-
-    @property
-    def is_on(self):
-        """Return true if the switch is on."""
-        return self.sensor_instance.is_on()
-
-    async def async_added_to_hass(self):
-        """Entity added to HA (at startup or when re-enabled)."""
-        await super().async_added_to_hass()
-        _LOGGER.debug("%s added to HA", self.name)
-        self.sensor_instance.set_enabled(True)
-        # request an update of this sensor
-        if self.device.is_online():
-            try:
-                await self.sensor_instance.async_update()
-            except ImouException as exception:
-                _LOGGER.error(exception.to_string())
-
-    async def async_will_remove_from_hass(self):
-        """Entity removed from HA (when disabled)."""
-        await super().async_added_to_hass()
-        _LOGGER.debug("%s removed from HA", self.name)
-        self.sensor_instance.set_enabled(False)