+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.1.2] (2022-10-03)
+
+### Added
+
+- All the switches are now made available. Only a subset are then enabled in HA by default.
+- Sensors' icons and default icon
+
+### Changed
+
+- Bump imouapi version: 0.1.5 → 0.2.0 and adapted the code accordingly
+- Introduced `ImouEntity` class for all the sensors derived subclasses
+
 ## [0.1.1] (2022-09-29)
 
 ### Changed